docs

+++ src/dewloosh/math/linalg/frontal/frontal.py
@@ -21,8 +21,8 @@ def frsolve(A: np.ndarray, B: np.ndarray, presc_bool: np.ndarray = None,
                 epath = np.arange(len(A))
         pre = presc_val is not None
         if not pre:
-            presc_bool = np.zeros((nEQ,), dtype=np.int32)
-            presc_val = np.zeros((nEQ,), dtype=np.float32)
+            presc_bool = np.zeros((nEQ,), dtype=np.int64)
+            presc_val = np.zeros((nEQ,), dtype=np.float64)
         lhs, rhs, eqpath, glob_to_front, glob_to_width = \
             frontal_sym_bulk_uniform(A, topology, B,
                                      presc_bool, presc_val, epath)
@@ -76,7 +76,7 @@ if __name__ == '__main__':
         cond = (condX | condY) & condZ
         nIDs = np.where(cond)[0]
         if solver == 'numpy':
-            p = np.array([1e12, 1e12, 1e12], dtype=np.float32)
+            p = np.array([1e12, 1e12, 1e12], dtype=np.float64)
             penalties = {nID: p for nID in nIDs}
         else:
             disps = {}
@@ -89,7 +89,7 @@ if __name__ == '__main__':
         if case == 0:
             nIDs = np.where(condZ)[0]
             fz = (-10/10000)*Lx*Ly/len(nIDs)
-            fp = np.array([0, 0, fz], dtype=np.float32)
+            fp = np.array([0, 0, fz], dtype=np.float64)
             loads = {nID: fp for nID in nIDs}
         elif case == 1:
             condZ = (abs(coord[:, 2] - Lz/2) < 0.001)
@@ -113,14 +113,14 @@ if __name__ == '__main__':
         nTOTV = len(points) * 3
         if solver == 'frontal':
             # essential boundary conditions
-            presc_bool = np.zeros(nTOTV, dtype=np.int32)
-            presc_val = np.zeros(nTOTV, dtype=np.float32)
+            presc_bool = np.zeros(nTOTV, dtype=np.int64)
+            presc_val = np.zeros(nTOTV, dtype=np.float64)
             for dID, dval in disps.items():
                 presc_val[dID] = dval
                 presc_bool[dID] = 1
 
             # natural boundary conditions
-            RHS = np.zeros((len(points), 3), dtype=np.float32)
+            RHS = np.zeros((len(points), 3), dtype=np.float64)
             for nID, f in loads.items():
                 RHS[nID] = f
 
@@ -150,10 +150,10 @@ if __name__ == '__main__':
                                     for ID in penalties.keys()])
             spdata = np.concatenate([v for v in penalties.values()])
             K += coo_matrix((spdata, (spind, spind)), shape=(nTOTV, nTOTV),
-                            dtype=np.float32)
+                            dtype=np.float64)
 
             # natural boundary conditions
-            RHS = np.zeros((len(points), 3), dtype=np.float32)
+            RHS = np.zeros((len(points), 3), dtype=np.float64)
             for nID, f in loads.items():
                 RHS[nID] = f
 
@@ -168,14 +168,14 @@ if __name__ == '__main__':
                 np.reshape(f - np.matmul(K.todense(), u), RHS.shape)
         elif solver == 'Gauss-Jordan':
             # essential boundary conditions
-            presc_bool = np.zeros(nTOTV, dtype=np.int32)
-            presc_val = np.zeros(nTOTV, dtype=np.float32)
+            presc_bool = np.zeros(nTOTV, dtype=np.int64)
+            presc_val = np.zeros(nTOTV, dtype=np.float64)
             for dID, dval in disps.items():
                 presc_val[dID] = dval
                 presc_bool[dID] = 1
 
             # natural boundary conditions
-            RHS = np.zeros((len(points), 3), dtype=np.float32)
+            RHS = np.zeros((len(points), 3), dtype=np.float64)
             for nID, f in loads.items():
                 RHS[nID] = f
 
@@ -189,14 +189,14 @@ if __name__ == '__main__':
             assembly.point_data['res'] = R.reshape(RHS.shape)
         elif solver == 'Jordan':
             # essential boundary conditions
-            presc_bool = np.zeros(nTOTV, dtype=np.int32)
-            presc_val = np.zeros(nTOTV, dtype=np.float32)
+            presc_bool = np.zeros(nTOTV, dtype=np.int64)
+            presc_val = np.zeros(nTOTV, dtype=np.float64)
             for dID, dval in disps.items():
                 presc_val[dID] = dval
                 presc_bool[dID] = 1
 
             # natural boundary conditions
-            RHS = np.zeros((len(points), 3), dtype=np.float32)
+            RHS = np.zeros((len(points), 3), dtype=np.float64)
             for nID, f in loads.items():
                 RHS[nID] = f
